+++ geo_primitives/src/direction_3d_transform_safe.rs
@@ -309,9 +309,9 @@ mod tests {
         let expected: Direction3D<f64> = Direction3D::positive_y();
 
         let tolerance = 1e-10_f64;
-        assert!((result.x() as f64 - expected.x() as f64).abs() < tolerance);
-        assert!((result.y() as f64 - expected.y() as f64).abs() < tolerance);
-        assert!((result.z() as f64 - expected.z() as f64).abs() < tolerance);
+        assert!((result.x() - expected.x()).abs() < tolerance);
+        assert!((result.y() - expected.y()).abs() < tolerance);
+        assert!((result.z() - expected.z()).abs() < tolerance);
     }
 
     #[test]
@@ -323,9 +323,9 @@ mod tests {
         let expected: Direction3D<f64> = Direction3D::positive_z();
 
         let tolerance = 1e-10_f64;
-        assert!((result.x() as f64 - expected.x() as f64).abs() < tolerance);
-        assert!((result.y() as f64 - expected.y() as f64).abs() < tolerance);
-        assert!((result.z() as f64 - expected.z() as f64).abs() < tolerance);
+        assert!((result.x() - expected.x()).abs() < tolerance);
+        assert!((result.y() - expected.y()).abs() < tolerance);
+        assert!((result.z() - expected.z()).abs() < tolerance);
     }
 
     #[test]
@@ -337,9 +337,9 @@ mod tests {
         let expected: Direction3D<f64> = Direction3D::positive_x();
 
         let tolerance = 1e-10_f64;
-        assert!((result.x() as f64 - expected.x() as f64).abs() < tolerance);
-        assert!((result.y() as f64 - expected.y() as f64).abs() < tolerance);
-        assert!((result.z() as f64 - expected.z() as f64).abs() < tolerance);
+        assert!((result.x() - expected.x()).abs() < tolerance);
+        assert!((result.y() - expected.y()).abs() < tolerance);
+        assert!((result.z() - expected.z()).abs() < tolerance);
     }
 
     #[test]
@@ -354,9 +354,9 @@ mod tests {
         let expected: Direction3D<f64> = Direction3D::negative_x();
 
         let tolerance = 1e-10_f64;
-        assert!((result.x() as f64 - expected.x() as f64).abs() < tolerance);
-        assert!((result.y() as f64 - expected.y() as f64).abs() < tolerance);
-        assert!((result.z() as f64 - expected.z() as f64).abs() < tolerance);
+        assert!((result.x() - expected.x()).abs() < tolerance);
+        assert!((result.y() - expected.y()).abs() < tolerance);
+        assert!((result.z() - expected.z()).abs() < tolerance);
     }
 
     #[test]
@@ -392,9 +392,9 @@ mod tests {
         let expected: Direction3D<f64> = Direction3D::positive_x(); // X成分は変わらない
 
         let tolerance = 1e-10_f64;
-        assert!((result.x() as f64 - expected.x() as f64).abs() < tolerance);
-        assert!((result.y() as f64 - expected.y() as f64).abs() < tolerance);
-        assert!((result.z() as f64 - expected.z() as f64).abs() < tolerance);
+        assert!((result.x() - expected.x()).abs() < tolerance);
+        assert!((result.y() - expected.y()).abs() < tolerance);
+        assert!((result.z() - expected.z()).abs() < tolerance);
     }
 
     #[test]
